clean clippy and check plus docs and some refactors

+++ benches/eig.rs
@@ -6,29 +6,46 @@ use ndarray_linalg::Eig;
 // use faer::prelude::*;
 // use faer::mat;
 
-fn eigen_ndarray(matrix: &Array2<f64>) -> Result<()> {
+/// Performs eigen decomposition using `ndarray-linalg`
+///
+/// # Arguments
+///
+/// * `matrix` - A reference to a 2D array of type `f32`
+///
+/// # Returns
+///
+/// A `Result` indicating success or failure.
+fn eigen_ndarray(matrix: &Array2<f32>) -> Result<()> {
     let _ = matrix.clone().eig().unwrap();
     Ok(())
 }
 
-fn eigen_nalgebra(matrix: &DMatrix<f64>) -> Result<()> {
+/// Performs eigen decomposition using `nalgebra`
+///
+/// # Arguments
+///
+/// * `matrix` - A reference to a `DMatrix` of type `f32`
+///
+/// # Returns
+///
+/// A `Result` indicating success or failure.
+fn eigen_nalgebra(matrix: &DMatrix<f32>) -> Result<()> {
     let _ = matrix.clone().symmetric_eigen();
     Ok(())
 }
 
-// fn eigen_faer(matrix: &Mat<f64>) -> Result<(), ()> {
-//     let _ = matrix.eigendecomposition::<c64>();
-//     Ok(())
-// }
-
+/// Benchmark functions for eigen decomposition.
+///
+/// # Arguments
+///
+/// * `c` - A mutable reference to a `Criterion` object for benchmarking.
 fn benchmarks(c: &mut Criterion) {
     let mut group = c.benchmark_group("Eigen Decomposition");
 
     // Example matrix for testing
     let size = 100;
-    let matrix_ndarray = Array2::<f64>::zeros((size, size));
-    let matrix_nalgebra = DMatrix::<f64>::zeros(size, size);
-    // let matrix_faer = Mat::zeros(size, size);
+    let matrix_ndarray = Array2::<f32>::zeros((size, size));
+    let matrix_nalgebra = DMatrix::<f32>::zeros(size, size);
 
     group.bench_function("Eigen Decomposition with ndarray-linalg", |b| {
         b.iter(|| eigen_ndarray(&matrix_ndarray))
@@ -36,10 +53,6 @@ fn benchmarks(c: &mut Criterion) {
     group.bench_function("Eigen Decomposition with nalgebra", |b| {
         b.iter(|| eigen_nalgebra(&matrix_nalgebra))
     });
-    // group.bench_function("Eigen Decomposition with faer", |b| {
-    //     b.iter(|| eigen_faer(&matrix_faer).unwrap())
-    // });
-
     group.finish();
 }
